+++ src/_stories/IdleTimeout.stories.js
@@ -1,10 +1,10 @@
 import { number, text, withKnobs } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
+import IdleTimeoutModal from 'components/IdleStatusMonitor'
 import _ from 'lodash/fp'
 import { Fragment, useEffect, useRef } from 'react'
 import { div, h, span } from 'react-hyperscript-helpers'
 import { ButtonPrimary } from 'src/components/common'
-import IdleTimeoutModal from 'src/components/IdleTimeoutModal'
 import Modal from 'src/components/Modal'
 import { getDynamic, listenDynamic, setDynamic } from 'src/libs/browser-storage'
 import { authStore } from 'src/libs/state'
