+++ src/pages/Main.js
@@ -1,5 +1,6 @@
 import 'src/libs/routes'
 
+import IdleStatusMonitor from 'components/IdleStatusMonitor'
 import { hot } from 'react-hot-loader/root'
 import { h } from 'react-hyperscript-helpers'
 import AuthContainer from 'src/components/AuthContainer'
@@ -7,7 +8,6 @@ import ConfigOverridesWarning from 'src/components/ConfigOverridesWarning'
 import ErrorWrapper from 'src/components/ErrorWrapper'
 import FirecloudNotification from 'src/components/FirecloudNotification'
 import FreeCreditsModal from 'src/components/FreeCreditsModal'
-import IdleStatusMonitor from 'src/components/IdleTimeoutModal'
 import ImportStatus from 'src/components/ImportStatus'
 import Notifications from 'src/components/Notifications'
 import { NpsSurvey } from 'src/components/NpsSurvey'
